refactor: move public logic to base class
