+++ pipefuser/model_executor/pipelines/register.py
@@ -1,9 +1,8 @@
-
 from typing import Dict, Type, Union
 from diffusers.pipelines.pipeline_utils import DiffusionPipeline
 
 from pipefuser.logger import init_logger
-from pipefuser.model_executor.pipelines import PipeFuserPipelineBaseWrapper
+from .base_pipeline import PipeFuserPipelineBaseWrapper
 
 logger = init_logger(__name__)
