+++ pipefuser/model_executor/schedulers/scheduling_dpmsolver_multistep.py
@@ -1,22 +1,21 @@
 from typing import Optional, Tuple, Union
+
+import torch
+import torch.distributed
+
+from diffusers.utils.torch_utils import randn_tensor
 from diffusers.schedulers.scheduling_dpmsolver_multistep import (
     DPMSolverMultistepScheduler,
     SchedulerOutput,
 )
-from diffusers.utils.torch_utils import randn_tensor
-import torch
-import torch.distributed
 
-from pipefuser.config import ParallelConfig, RuntimeConfig
 from pipefuser.distributed import (
     get_pipeline_parallel_world_size,
+    get_sequence_parallel_world_size,
+    get_runtime_state,
 )
-from pipefuser.distributed.parallel_state import get_sequence_parallel_world_size
-from pipefuser.distributed.runtime_state import get_runtime_state
-from pipefuser.model_executor.schedulers import (
-    PipeFuserSchedulerWrappersRegister,
-)
-from pipefuser.model_executor.schedulers.base_scheduler import PipeFuserSchedulerBaseWrapper
+from .register import PipeFuserSchedulerWrappersRegister
+from .base_scheduler import PipeFuserSchedulerBaseWrapper
 
 
 @PipeFuserSchedulerWrappersRegister.register(DPMSolverMultistepScheduler)
@@ -62,12 +61,11 @@ class PipeFuserDPMSolverMultistepSchedulerWrapper(PipeFuserSchedulerBaseWrapper)
             or self.config.final_sigmas_type == "zero"
         )
         lower_order_second = (
-            (self.step_index == len(self.timesteps) - 2)
-            and self.config.lower_order_final
-            and len(self.timesteps) < 15
+            (self.step_index == len(self.timesteps) - 2) and self.config.lower_order_final and len(self.timesteps) < 15
         )
 
         model_output = self.convert_model_output(model_output, sample=sample)
+
         if (
             get_runtime_state().patch_mode
             and get_runtime_state().pipeline_patch_idx == 0
@@ -87,8 +85,6 @@ class PipeFuserDPMSolverMultistepSchedulerWrapper(PipeFuserSchedulerBaseWrapper)
             for i in range(self.config.solver_order - 1):
                 self.model_outputs[i] = self.model_outputs[i + 1]
 
-        _, _, c, _ = model_output.shape
-
         if get_runtime_state().patch_mode and get_runtime_state().pipeline_patch_idx == 0:
             assert len(self.model_outputs) >= 2
             self.model_outputs[-1] = torch.zeros_like(self.model_outputs[-2])
@@ -105,15 +101,9 @@ class PipeFuserDPMSolverMultistepSchedulerWrapper(PipeFuserSchedulerBaseWrapper)
 
         # Upcast to avoid precision issues when computing prev_sample
         sample = sample.to(torch.float32)
-        if (
-            self.config.algorithm_type in ["sde-dpmsolver", "sde-dpmsolver++"]
-            and variance_noise is None
-        ):
+        if self.config.algorithm_type in ["sde-dpmsolver", "sde-dpmsolver++"] and variance_noise is None:
             noise = randn_tensor(
-                model_output.shape,
-                generator=generator,
-                device=model_output.device,
-                dtype=torch.float32,
+                model_output.shape, generator=generator, device=model_output.device, dtype=torch.float32,
             )
         elif self.config.algorithm_type in ["sde-dpmsolver", "sde-dpmsolver++"]:
             noise = variance_noise.to(device=model_output.device, dtype=torch.float32)
@@ -137,26 +127,12 @@ class PipeFuserDPMSolverMultistepSchedulerWrapper(PipeFuserSchedulerBaseWrapper)
         else:
             model_outputs = self.model_outputs
 
-        if (
-            self.config.solver_order == 1
-            or self.lower_order_nums < 1
-            or lower_order_final
-        ):
-            prev_sample = self.dpm_solver_first_order_update(
-                model_output, sample=sample, noise=noise
-            )
-        elif (
-            self.config.solver_order == 2
-            or self.lower_order_nums < 2
-            or lower_order_second
-        ):
-            prev_sample = self.multistep_dpm_solver_second_order_update(
-                model_outputs, sample=sample, noise=noise
-            )
+        if self.config.solver_order == 1 or self.lower_order_nums < 1 or lower_order_final:
+            prev_sample = self.dpm_solver_first_order_update(model_output, sample=sample, noise=noise)
+        elif self.config.solver_order == 2 or self.lower_order_nums < 2 or lower_order_second:
+            prev_sample = self.multistep_dpm_solver_second_order_update(model_outputs, sample=sample, noise=noise)
         else:
-            prev_sample = self.multistep_dpm_solver_third_order_update(
-                model_outputs, sample=sample
-            )
+            prev_sample = self.multistep_dpm_solver_third_order_update(model_outputs, sample=sample)
 
         if self.lower_order_nums < self.config.solver_order:
             self.lower_order_nums += 1
@@ -165,6 +141,7 @@ class PipeFuserDPMSolverMultistepSchedulerWrapper(PipeFuserSchedulerBaseWrapper)
         prev_sample = prev_sample.to(model_output.dtype)
 
         # upon completion increase step index by one
+        #* increase step index only when the last pipeline patch is done (or not in patch mode)
         if (
             not get_runtime_state().patch_mode
             or get_runtime_state().pipeline_patch_idx == get_runtime_state().num_pipeline_patch - 1
